Cast to floats when comparing episode numbers to account for "N.5" episodes.

+++ aniflow/common.py
@@ -35,7 +35,7 @@ class Episode:
         elif not self.episode_number:
             return True
 
-        return int(self.episode_number) is int(self.anilist_entry.episode_count)
+        return float(self.episode_number) is float(self.anilist_entry.episode_count)
 
     def _get_file_name(self):
         """Returns the file name after removing any directory paths"""
